==> 1-daily-data-update.ipynb <==
# Daily Fantasy Model Data Update

Incremental daily updates to season, player, boxscore, lineup and daily fantasy salary information from mysportsfeeds.

The dataload is dependent on the following external (mysportsfeeds) APIs:
- season games api:  https://api.mysportsfeeds.com/v2.0/pull/nba/{season}/games.{format}
- boxscore api: https://api.mysportsfeeds.com/v2.0/pull/nba/{season}/date/{date}/player_gamelogs.{format}
- game lineup api: https://api.mysportsfeeds.com/v2.0/pull/nba/{season}/games/{game}/lineup.{format}
- play-by-play api: https://api.mysportsfeeds.com/v2.0/pull/nba/{season}/games/{game}/playbyplay.{format}
- players api: https://api.mysportsfeeds.com/v2.0/pull/nba/players.{format}

%load_ext autoreload
%autoreload 2

from time import sleep
from datetime import date
import datetime
import os

import pandas as pd
import numpy as np
import pymongo
from ohmysportsfeedspy import MySportsFeeds
from tqdm.notebook import tqdm as tqdmm

import dataloader as dl

MSF_KEY = os.environ.get("MSF_KEY")
NBA_MONGO_KEY = os.environ.get("NBA_MONGO_KEY")

pd.options.display.max_columns = 999

msf = MySportsFeeds(version="2.0")
msf.authenticate(MSF_KEY, "MYSPORTSFEEDS")

client = pymongo.MongoClient(f"mongodb+srv://nba-fantasy:{NBA_MONGO_KEY}@nba-fantasy-hu5fx.gcp.mongodb.net/test?retryWrites=true&w=majority")

## Date range
Specify date range for updating nba stats

start = datetime.datetime.strptime("20200115", "%Y%m%d")
end = datetime.datetime.strptime("20200116", "%Y%m%d")

## Season games

game_parser = dl.GameDataLoader(client)

season = "2019-2020-regular"
payload = msf.msf_get_data(league='nba', feed='seasonal_games', season=season, format='json')
payload["season"] = season
parsed_games = game_parser.parse(payload)
for game in tqdmm(parsed_games):
    game_parser.upload(game)

## Players

player_parser = dl.PlayerDataLoader(client)

payload = msf.msf_get_data(league='nba', feed='players', format='json')
parsed_players = player_parser.parse(payload)
for player in tqdmm(parsed_players):
    player_parser.upload(player)

## DFS Salaries

season = "2019-2020-regular"

fanduel_parser = dl.FanduelDataLoader(client)
date_generated = [(start + datetime.timedelta(days=x)).strftime("%Y%m%d") for x in range(0, (end-start).days + 1)]

for game_date in date_generated:
    payload = msf.msf_get_data(league='nba', feed='daily_dfs', season=season, date=game_date, format='json')
    try:
        parsed_dfs = fanduel_parser.parse(payload)
        for dfs in tqdmm(parsed_dfs):
            fanduel_parser.upload(dfs)
    except KeyError as e:
        print(f"Unable to parse games for date {game_date}. Error: ", e)

## Player boxscores

gamelog_parser = dl.GameLogDataLoader(client)
date_generated = [(start + datetime.timedelta(days=x)).strftime("%Y%m%d") for x in range(0, (end-start).days + 1)]

for game_date in date_generated:
    payload = msf.msf_get_data(league="nba", feed="daily_player_gamelogs", season=season, date=game_date, force=False, format="json")
    parsed_gamelogs = gamelog_parser.parse(payload)
    for game in tqdmm(parsed_gamelogs):
        gamelog_parser.upload(game)

## Game Lineup

date_generated = [(start + datetime.timedelta(days=x)).strftime("%Y%m%d") for x in range(0, (end-start).days + 1)]

for date in date_generated:
    payload = msf.msf_get_data(league="nba", feed="daily_games", season=season, date=date, force=False, format="json")
    game_ids = list(map(lambda x: x["schedule"]["id"], payload["games"]))

    lineup_parser = dl.LineupDataLoader(client)

    lineups_uploaded = []
    lineups_failed = []

    for game_id in tqdmm(game_ids):
        try:
            payload = msf.msf_get_data(league="nba", feed="game_lineup", season=season, game=game_id, force=False, format="json")
            parsed_lineups = lineup_parser.parse(payload)
            lineup_parser.upload(parsed_lineups)
            lineups_uploaded.append(game_id)
        except Exception as e:
            print(f"Upload for {game_id} failed with error `{e}`")
            lineups_failed.append(game_id)

## Are there any completed games that do not have a lineup or boxscore?

cursor = client["nbafantasy"]["game"].aggregate([
    {
        '$addFields': {
            'game_date': {
                '$dateToString': {
                    'date': {
                        '$add': [
                            {
                                '$toDate': '$schedule.startTime'
                            }, -18000000
                        ]
                    },
                    'format': '%Y%m%d'
                }
            }
        }
    },
    {
        "$match": {
            "schedule.playedStatus": {"$eq": "COMPLETED"}
        }
    },
    {
        "$project": {
            "_id": 0,
            "season": 1,
            "game_date": 1,
            "game_id": "$schedule.id"
        }
    }
])
completed_games = [(x["season"], x["game_date"], x["game_id"]) for x in cursor]

games_with_boxscores = {
    x["game"]["id"] for x in gamelog_parser.download({}, ["game.id"])
}

games_with_completed_lineups = {
    (x["_id"])
    for x in lineup_parser.download(
        {"game.playedStatus": {"$eq": "COMPLETED"}},
        ["_id"]
    )
}

# games without boxscores
games_without_boxscores = list(filter(lambda x: x[2] not in games_with_boxscores, completed_games)); games_without_boxscores

# games without lineups
games_without_lineups = list(filter(lambda x: x[2] not in games_with_completed_lineups, completed_games)); games_without_lineups

# download missing boxscores
gamelog_parser = dl.GameLogDataLoader(client)
for missing_date in tqdmm(games_without_boxscores):
    payload = msf.msf_get_data(
        league="nba",
        feed="daily_player_gamelogs",
        season=missing_date[0],
        date=missing_date[1], 
        game=missing_date[2],
        force=False,
        format="json"
    )
    parsed_gamelogs = gamelog_parser.parse(payload)
    for player_game in parsed_gamelogs:
        gamelog_parser.upload(player_game)
